NOT SURE IF I SUBMITTED CORRECTLY BEFORE.  IF NOT HERE IS THE OTHER WAY I WAS THINKING ABOUT SUBMITTING IT
I re-did the XML because there were no cell id's and everything was set
to text even though we needed images.  Also the buttons had set width
and height, when they should have gotten a new table row and laid out
like that.  The XML needs more work to get the images to fit correctly
(try running the app, and you'll see what I mean.  It only displays the
first 6 images.)  The Main Activity is running OK, however there is a
glitch with the buttons that prevents it from running, and the setAlpha
method is not working how I intended.  I'm going to be around the CS
Lounge at 7pm to see if the TA can help, if anyone wants to drop by to
lend a hand, especially with XML, it would be greatly appreciated.
There is still a good amount to do in Main Activity (I can't test it
efficiently until I fix up the XML).